refactor(dashboard): split agent-detail.ts into sub-components
Extract 474-line agent-detail.ts into focused modules:
- agent-detail-state.ts (190 lines): signals, selectors, data fetching, helpers
- agent-detail-timeline.ts (65 lines): AgentTimelineSection component
- agent-detail-journal.ts (30 lines): AgentJournalStream component
- agent-detail-worker.ts (48 lines): AgentWorkerBrief component
- agent-detail.ts (199 lines): main AgentDetailOverlay, re-exports public API

All files under 300 lines. External imports unchanged (openAgentDetail,
closeAgentDetail, selectedAgentName still exported from agent-detail.ts).

Closes #1518
