+++ dashboard/src/components/agent-detail-timeline.ts
@@ -1,0 +1,65 @@
+// Agent detail timeline section — activity timeline with event summary
+
+import { html } from 'htm/preact'
+import { Card } from './common/card'
+import { TimeAgo } from './common/time-ago'
+import { agentTimeline, compactCopy } from './agent-detail-state'
+import type { AgentTimelineEvent } from '../api'
+
+function timelineEventIcon(type: string): string {
+  if (type === 'joined') return 'J'
+  if (type.startsWith('task_')) return 'T'
+  if (type === 'broadcast') return 'M'
+  return 'E'
+}
+
+function timelineEventLabel(type: string): string {
+  switch (type) {
+    case 'joined': return '참가'
+    case 'task_claimed': return '태스크 수임'
+    case 'task_started': return '태스크 시작'
+    case 'task_completed': return '태스크 완료'
+    case 'task_cancelled': return '태스크 취소'
+    case 'broadcast': return '브로드캐스트'
+    default: return type
+  }
+}
+
+export function AgentTimelineSection() {
+  const timeline = agentTimeline.value
+  if (!timeline) return null
+
+  const events = timeline.events ?? []
+  const summary = timeline.summary
+
+  return html`
+    <${Card} title="활동 타임라인 (${summary?.total_events ?? 0} events)">
+      ${summary ? html`
+        <div class="agent-timeline-summary">
+          ${summary.tasks_completed > 0 ? html`<span class="pill">완료 ${summary.tasks_completed}</span>` : null}
+          ${summary.tasks_claimed > 0 ? html`<span class="pill">수임 ${summary.tasks_claimed}</span>` : null}
+          ${summary.messages_sent > 0 ? html`<span class="pill">메시지 ${summary.messages_sent}</span>` : null}
+          ${summary.active_duration_minutes > 0 ? html`<span class="pill">${Math.round(summary.active_duration_minutes)}분 활동</span>` : null}
+        </div>
+      ` : null}
+      ${events.length === 0
+        ? html`<div class="empty-state">타임라인 이벤트 없음</div>`
+        : html`
+            <div class="agent-timeline-list">
+              ${events.map((evt: AgentTimelineEvent, idx: number) => {
+                const detail = evt.detail as Record<string, string | undefined>
+                const title = detail.title ?? detail.content ?? ''
+                return html`
+                  <div class="agent-timeline-event" key=${idx}>
+                    <span class="agent-journal-kind">${timelineEventIcon(evt.type)}</span>
+                    <span class="agent-timeline-type">${timelineEventLabel(evt.type)}</span>
+                    ${title ? html`<span class="agent-timeline-detail">${compactCopy(title, 80)}</span>` : null}
+                    ${evt.ts ? html`<${TimeAgo} timestamp=${evt.ts} />` : null}
+                  </div>
+                `
+              })}
+            </div>
+          `}
+    <//>
+  `
+}
